refactor: extract function to utils
